restored old casts (there was a reason for them)
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@14646 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/msw/treectrl.cpp
@@ -1514,7 +1514,7 @@ wxTreeItemId wxTreeCtrl::AddRoot(const wxString& text,
         return TVI_ROOT;
     }
 
-    return DoInsertItem(wxTreeItemId((WXHTREEITEM) 0), (WXHTREEITEM) 0,
+    return DoInsertItem(wxTreeItemId((long)(WXHTREEITEM) 0), (long)(WXHTREEITEM) 0,
                         text, image, selectedImage, data);
 }
